// typescript/user-interface/packages/ui-state/src/ts/app/query/types.ts
export enum AsyncActionStatus {
  /** the action is pending */
  pending = 'pending',
  /** the action has been fulfilled and completed */
  fulfilled = 'fulfilled',
  /** the action has been rejected or failed */
  rejected = 'rejected'
}

/**
 * Defines the base history entry object for an async fetch request.
 */
export interface AsyncFetchHistoryEntry<T> {
  /** the async request arguments */
  arg: T;
  /** the async action request status */
  status: AsyncActionStatus;
  /** the number of attempts the query has been issued */
  attempts: number;
  /** the time in epoch seconds of when the last attempt was completed */
  time: number;
  /**
   * The serialized error from the rejected promise, but excluding any non-serializable parameters.
   * See @link https://redux-toolkit.js.org/api/createAsyncThunk#handling-thunk-errors
   */
  error?: SerializedError | SerializedAxiosError;
}

/**
 * Defines the base history object for an async fetch request.
 */
export type AsyncFetchHistory<T> = Record<
  string /* unique string identifier */,
  Record<string /* request id */, AsyncFetchHistoryEntry<T>>
>;

interface BaseAsyncFetchRequest<T> {
  /** the number of pending (active) fetch requests */
  pending: number;
  /** the number of fulfilled (completed) fetch requests */
  fulfilled: number;
  /** the number of rejected fetch requests */
  rejected: number;
  /** true if fetching is active or loading; false otherwise */
  isLoading: boolean;
  /** true if an error occurred while fetching; false otherwise */
  isError: boolean;
  /** the data of the result */
  data?: T;
  /** the serializable errors */
  error?: (SerializedError | SerializedAxiosError)[];
}

/**
 * Defines the base result object for an async fetch request.
 */
export type AsyncFetchResult<T> =
  | Override<
      BaseAsyncFetchRequest<T>,
      | Omit<BaseAsyncFetchRequest<T>, 'error'>
      | ({
          isLoading: false;
          isError: false;
        } & Required<Pick<BaseAsyncFetchRequest<T>, 'data'>>)
    >
  | Override<
      BaseAsyncFetchRequest<T>,
      | Omit<BaseAsyncFetchRequest<T>, 'data'>
      | {
          isLoading: true;
          isError: false;
        }
    >
  | Override<
      BaseAsyncFetchRequest<T>,
      | Omit<BaseAsyncFetchRequest<T>, 'data'>
      | ({
          isLoading: false;
          isError: true;
        } & Required<Pick<BaseAsyncFetchRequest<T>, 'error'>>)
    >;